requirements.txt
langchain
langchain-openai
langchain-community
openai
faiss-cpu
tiktoken
python-dotenv
